+++ pkg/libcontainer/namespaces/create.go
@@ -1,4 +1,4 @@
-package nsinit
+package namespaces
 
 import (
 	"os"
