+++ pkg/libcontainer/namespaces/pid.go
@@ -1,4 +1,4 @@
-package nsinit
+package namespaces
 
 import (
 	"fmt"
