+++ pkg/libcontainer/namespaces/sync_pipe.go
@@ -1,4 +1,4 @@
-package nsinit
+package namespaces
 
 import (
 	"encoding/json"
